FIX: Align ordering with Magma for orders 3^7,5^7,7^7,11^7
The permutations applied had been wrongly implemented and were based on
wrong data. New permutations, based on explicitly comparing with Magma data
were used to determine correct permutations.

The test for the permutations in `ordering.tst` was changed, as the old test
relied on the permutations having small support, which is not true any
longer.

This will fix #38

Removed the old tests as they required that the permutations were of small
support. (They also were wrong)

Fixed hashmark and added tests

Added documentation for the new ordering.